+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/parsing/parsed_method_node.rb
@@ -13,6 +13,13 @@ class RubyLsp::Ree::ParsedMethodNode
     @method_node.name
   end
 
+  def param_names
+    @method_node.parameters.requireds.map(&:name) + 
+    @method_node.parameters.keywords.map(&:name) + 
+    [@method_node.parameters.rest&.name] + 
+    [@method_node.parameters.keyword_rest&.name]
+  end
+
   def has_contract?
     !!@contract_node
   end