+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/ree_object_finder.rb
@@ -1,3 +1,4 @@
+require 'delegate'
 require_relative "utils/ree_lsp_utils"
 
 module RubyLsp
@@ -8,6 +9,20 @@ module RubyLsp
       MAX_LIMIT = 1000
 
       REE_OBJECT_STRING = 'ree_object'
+
+      class ReeObjectDecorator < SimpleDelegator
+        include RubyLsp::Ree::ReeLspUtils
+
+        def object_package
+          return @package_name if @package_name
+
+          package_name_from_uri(uri)
+        end
+
+        def set_package!(package_name)
+          @package_name = package_name
+        end
+      end
 
       def initialize(index)
         @index = index
@@ -41,6 +56,14 @@ module RubyLsp
         objects_by_name.detect{ _1.comments.to_s.lines.first&.chomp == REE_OBJECT_STRING }
       end
 
+      def find_objects(name)
+        objects_by_name = @index[name]
+        return unless objects_by_name
+
+        ree_objects = objects_by_name.select{ _1.comments.to_s.lines.first&.chomp == REE_OBJECT_STRING }
+        decorate_objects(ree_objects)
+      end
+
       def find_object_for_package(name, package_name)
         objects_by_name = @index[name]
         return unless objects_by_name
@@ -64,6 +87,12 @@ module RubyLsp
 
       def object_documentation(ree_object)
         ree_object.comments.lines[2..-1].join("\n").chomp
+      end
+
+      private 
+      
+      def decorate_objects(ree_objects)
+        ree_objects.map{ ReeObjectDecorator.new(_1) }
       end
     end
   end