+++ ruby-lsp-ree/spec/formatter/missing_imports_spec.rb
@@ -165,6 +165,29 @@ RSpec.describe "RubyLsp::Ree::ReeFormatter" do
       expect(result.lines[2].strip).to eq('link :seconds_ago')
       expect(result.lines[3].strip).to eq('end')
     end
+
+    it "adds import from nested method call" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+  
+          def call(arg1)
+            local_method(
+              local_method([1, seconds_ago])
+            )
+          end
+
+          def local_method(arg1)
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting(sample_file_uri, ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class do')
+      expect(result.lines[2].strip).to eq('link :seconds_ago')
+      expect(result.lines[3].strip).to eq('end')
+    end
   end
 
   context "bean method calls" do
@@ -194,6 +217,24 @@ RSpec.describe "RubyLsp::Ree::ReeFormatter" do
           SOME_CONST = seconds_ago.call_method
   
           def call(arg1)
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting(sample_file_uri, ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class do')
+      expect(result.lines[2].strip).to eq('link :seconds_ago')
+      expect(result.lines[3].strip).to eq('end')
+    end
+
+    it "adds import link for calls from if predicate" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+          
+          def call(arg1)
+            return false if !seconds_ago.call_method
           end
         end
       RUBY
@@ -263,6 +304,70 @@ RSpec.describe "RubyLsp::Ree::ReeFormatter" do
       expect(result.lines[1].strip).to eq('fn :some_class')
     end
 
+    it "doesn't add import if receiver is method argument" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+          
+          def call(seconds_ago)
+            seconds_ago.call_method
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting(sample_file_uri, ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class')
+    end
+
+    it "doesn't add import if receiver is keyword argument" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+          
+          def call(seconds_ago:)
+            seconds_ago.call_method
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting(sample_file_uri, ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class')
+    end
+
+    it "doesn't add import if receiver is splat args" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+          
+          def call(*seconds_ago)
+            seconds_ago.call_method
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting(sample_file_uri, ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class')
+    end
+
+    it "doesn't add import if receiver is kwargs" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+          
+          def call(**seconds_ago)
+            seconds_ago.call_method
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting(sample_file_uri, ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class')
+    end
+
     it "doesn't add import if bean has receiver" do
       source =  <<~RUBY
         class SamplePackage::SomeClass
@@ -277,6 +382,99 @@ RSpec.describe "RubyLsp::Ree::ReeFormatter" do
       result = subject.run_formatting(sample_file_uri, ruby_document(source))
   
       expect(result.lines[1].strip).to eq('fn :some_class')
+    end
+
+    it "adds import from nested bean method call" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+  
+          def call(arg1)
+            local_method(
+              seconds_ago.by_second(1).active,
+              only: [:id]
+            ).first
+          end
+
+          def local_method(arg1, fields)
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting(sample_file_uri, ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class do')
+      expect(result.lines[2].strip).to eq('link :seconds_ago')
+      expect(result.lines[3].strip).to eq('end')
+    end
+  end
+
+  context "multiple found objects" do
+    before :each do
+      with_server('') do |server, uri|
+        index_fn(server, 'duplicated_fn', 'package1')
+        index_fn(server, 'duplicated_fn', 'sample_package')
+        index_fn(server, 'duplicated_fn', 'package2')
+        
+        index_fn(server, 'do_something_with_my_date', 'ree_date')
+        index_fn(server, 'do_something_with_my_date', 'ree_datetime')
+
+        @index = server.global_state.index 
+      end
+    end
+
+    it "adds import from same package by default" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+  
+          def call(arg1)
+            duplicated_fn
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting(sample_file_uri, ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class do')
+      expect(result.lines[2].strip).to eq('link :duplicated_fn')
+      expect(result.lines[3].strip).to eq('end')
+    end
+
+    it "chooses datetime between ree_datetime and ree_date" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+  
+          def call(arg1)
+            do_something_with_my_date
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting(sample_file_uri, ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class do')
+      expect(result.lines[2].strip).to eq('link :do_something_with_my_date, from: :ree_datetime')
+      expect(result.lines[3].strip).to eq('end')
+    end
+
+    it "adds import with placeholder if multiple objects found" do
+      source =  <<~RUBY
+        class SamplePackage::SomeClass
+          fn :some_class
+  
+          def call(arg1)
+            duplicated_fn
+          end
+        end
+      RUBY
+  
+      result = subject.run_formatting('', ruby_document(source))
+  
+      expect(result.lines[1].strip).to eq('fn :some_class do')
+      expect(result.lines[2].strip).to eq('link :duplicated_fn, from: FILL_PACKAGE')
+      expect(result.lines[3].strip).to eq('end')
     end
   end
 end